feat: refactoring app code into internal packages
