+++ internal/dbqueries/query_loader.go
@@ -1,14 +1,13 @@
-package main
+package dbqueries
 
 import (
 	"context"
 	"database/sql"
 
 	_ "github.com/mattn/go-sqlite3"
-	"github.com/mr55p-dev/pagemail/internal/dbqueries"
 )
 
-func mustGetDb(ctx context.Context, path string) (*dbqueries.Queries, func() error) {
+func MustGetQueries(ctx context.Context, path string) (*Queries, func() error) {
 	conn, err := sql.Open("sqlite3", path)
 	if err != nil {
 		panic(err)
@@ -17,5 +16,5 @@ func mustGetDb(ctx context.Context, path string) (*dbqueries.Queries, func() err
 	if err != nil {
 		panic(err)
 	}
-	return dbqueries.New(conn), conn.Close
+	return New(conn), conn.Close
 }
