+++ internal/request/bind_test.go
@@ -1,4 +1,4 @@
-package main
+package request
 
 import (
 	"log/slog"
