+++ internal/router/config.go
@@ -1,4 +1,4 @@
-package main
+package router
 
 import (
 	"log/slog"
