+++ internal/router/env.go
@@ -1,0 +1,15 @@
+package router
+
+type Env string
+type ContentType string
+
+const (
+	ENV_DEV Env = "dev"
+	ENV_STG Env = "stg"
+	ENV_PRD Env = "prd"
+
+	CONTENT_ANY   ContentType = "*/*"
+	CONTENT_HTML  ContentType = "text/html"
+	CONTENT_JSON  ContentType = "text/json"
+	CONTENT_PLAIN ContentType = "text/plain"
+)
