+++ internal/router/router.go
@@ -1,17 +1,12 @@
-package main
+package router
 
 import (
 	"context"
 	"fmt"
 	"io/fs"
-	"log/slog"
 	"net/http"
-	"os"
-	"strings"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/config"
-	"github.com/mr55p-dev/gonk"
 	"github.com/mr55p-dev/pagemail/internal/assets"
 	"github.com/mr55p-dev/pagemail/internal/auth"
 	"github.com/mr55p-dev/pagemail/internal/dbqueries"
@@ -20,117 +15,18 @@ import (
 	"github.com/mr55p-dev/pagemail/internal/middlewares"
 )
 
-var logger = logging.NewLogger("routes")
-
-type Env string
-type ContentType string
-
-const (
-	ENV_DEV Env = "dev"
-	ENV_STG Env = "stg"
-	ENV_PRD Env = "prd"
-
-	CONTENT_ANY   ContentType = "*/*"
-	CONTENT_HTML  ContentType = "text/html"
-	CONTENT_JSON  ContentType = "text/json"
-	CONTENT_PLAIN ContentType = "text/plain"
-)
+var logger = logging.NewLogger("router")
 
 type Router struct {
 	DBClient   *dbqueries.Queries
 	Authorizer *auth.Authorizer
+	Mux        http.Handler
 }
 
-type AccountData struct {
-	Subscribed string `form:"email-list"`
-}
-
-func HandleMethod(method string, handler http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.Method == method {
-			handler.ServeHTTP(w, r)
-		} else {
-			http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
-		}
-	})
-}
-
-func HandleMethods(methods map[string]http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if handler, ok := methods[r.Method]; ok {
-			handler.ServeHTTP(w, r)
-		} else {
-			http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
-		}
-	})
-}
-
-func ParseLogLvl(level string) slog.Level {
-	switch strings.ToUpper(level) {
-	case "DEBUG":
-		return slog.LevelDebug
-	case "ERROR":
-		return slog.LevelError
-	case "WARN":
-		return slog.LevelWarn
-	case "INFO":
-		return slog.LevelInfo
-	default:
-		return slog.LevelInfo
-	}
-}
-
-func getConfig() *AppConfig {
-	cfg := new(AppConfig)
-	err := gonk.LoadConfig(
-		cfg,
-		gonk.FileLoader("pagemail.yaml", true),
-		gonk.EnvironmentLoader("pm"),
-	)
-	if err != nil {
-		panic(err)
-	}
-	return cfg
-}
-
-func getAwsConfig(ctx context.Context) aws.Config {
-	awsCfg, err := config.LoadDefaultConfig(ctx)
-	if err != nil {
-		panic(err)
-	}
-	return awsCfg
-}
-
-func getLogger(cfg *AppConfig) *logging.Logger {
-	var lvl = slog.LevelInfo
-	switch strings.ToUpper(cfg.LogLevel) {
-	case "DEBUG":
-		lvl = slog.LevelDebug
-	case "ERROR":
-		lvl = slog.LevelError
-	case "WARN":
-		lvl = slog.LevelWarn
-	case "INFO":
-		lvl = slog.LevelInfo
-	}
-	handler := slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{
-		Level: lvl,
-	})
-	logger := logging.NewLogger("main")
-	logging.SetHandler(handler)
-	return logger
-}
-
-func main() {
-	// Load config
-	ctx := context.Background()
-	cfg := getConfig()
-	awsCfg := getAwsConfig(ctx)
-	logger := getLogger(cfg)
-
+func New(ctx context.Context, cfg *AppConfig, awsCfg aws.Config) (*Router, error) {
 	// Start the clients
 	logger.DebugCtx(ctx, "Setting up db client")
-	dbClient, dbClose := mustGetDb(ctx, cfg.DBPath)
+	dbClient, dbClose := dbqueries.MustGetQueries(ctx, cfg.DBPath)
 	defer func() {
 		_ = dbClose()
 	}()
@@ -216,9 +112,6 @@ func main() {
 		middlewares.RequestLogger,
 		middlewares.GetUserLoader(authClient, dbClient),
 	))
-
-	logger.Info("Starting http server", "config", cfg)
-	if err := http.ListenAndServe(cfg.Host, mux); err != nil {
-		panic(err)
-	}
+	s.Mux = mux
+	return s, nil
 }
