+++ test/test_integration.go
@@ -1,0 +1,10 @@
+//go:build integration
+
+package test
+
+import "testing"
+
+func TestServer(t *testing.T) {
+	// test server
+	// server :=
+}
